requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pymongo==4.3.3
Jinja2==3.1.2
python-dateutil==2.8.2
